+++ examples/with_auto_reauth.rs
@@ -45,19 +45,19 @@ async fn auth_code_do_things(spotify: &AuthCodeSpotify) {
 async fn client_creds_do_things(spotify: &ClientCredsSpotify) {
     // Running the requests
     let birdy_uri = AlbumId::from_uri("spotify:album:0sNOF9WDwhWunNAHPD3Baj").unwrap();
-    let albums = spotify.album(birdy_uri).await;
+    let albums = spotify.album(birdy_uri, None).await;
     println!("Get albums: {}", albums.unwrap().id);
 }
 
 async fn expire_token<S: BaseClient>(spotify: &S) {
     let token_mutex = spotify.get_token();
     let mut token = token_mutex.lock().await.unwrap();
-    let mut token = token.as_mut().expect("Token can't be empty as this point");
+    let token = token.as_mut().expect("Token can't be empty as this point");
     // In a regular case, the token would expire with time. Here we just do
     // it manually.
-    let now = Utc::now().checked_sub_signed(Duration::seconds(10));
+    let now = Utc::now().checked_sub_signed(Duration::try_seconds(10).unwrap());
     token.expires_at = now;
-    token.expires_in = Duration::seconds(0);
+    token.expires_in = Duration::try_seconds(0).unwrap();
     // We also use a garbage access token to make sure it's actually
     // refreshed.
     token.access_token = "garbage".to_owned();
@@ -113,7 +113,6 @@ async fn main() {
 
     // Enabling automatic token refreshing in the config
     let config = Config {
-        token_refreshing: true,
         ..Default::default()
     };