+++ src/clients/base.rs
@@ -9,7 +9,7 @@ use crate::{
     model::*,
     sync::Mutex,
     util::build_map,
-    ClientResult, Config, Credentials, Token,
+    ClientError, ClientResult, Config, Credentials, Token,
 };
 
 use std::{collections::HashMap, fmt, ops::Not, sync::Arc};
@@ -21,7 +21,8 @@ use serde_json::Value;
 /// This trait implements the basic endpoints from the Spotify API that may be
 /// accessed without user authorization, including parts of the authentication
 /// flow that are shared, and the endpoints.
-#[maybe_async]
+#[cfg_attr(target_arch = "wasm32", maybe_async(?Send))]
+#[cfg_attr(not(target_arch = "wasm32"), maybe_async)]
 pub trait BaseClient
 where
     Self: Send + Sync + Default + Clone + fmt::Debug,
@@ -93,18 +94,17 @@ where
     /// Since this is accessed by authenticated requests always, it's where the
     /// automatic reauthentication takes place, if enabled.
     #[doc(hidden)]
-    async fn auth_headers(&self) -> Headers {
-        self.auto_reauth()
-            .await
-            .expect("Failed to re-authenticate automatically, please authenticate");
+    async fn auth_headers(&self) -> ClientResult<Headers> {
+        self.auto_reauth().await?;
 
-        self.get_token()
+        Ok(self
+            .get_token()
             .lock()
             .await
-            .expect("Failed to acquire lock")
+            .unwrap()
             .as_ref()
-            .expect("RSpotify not authenticated")
-            .auth_headers()
+            .ok_or(ClientError::InvalidToken)?
+            .auth_headers())
     }
 
     // HTTP-related methods for the Spotify client. They wrap up the basic HTTP
@@ -116,7 +116,7 @@ where
     #[inline]
     async fn api_get(&self, url: &str, payload: &Query<'_>) -> ClientResult<String> {
         let url = self.api_url(url);
-        let headers = self.auth_headers().await;
+        let headers = self.auth_headers().await?;
         Ok(self.get_http().get(&url, Some(&headers), payload).await?)
     }
 
@@ -126,7 +126,7 @@ where
     #[inline]
     async fn api_post(&self, url: &str, payload: &Value) -> ClientResult<String> {
         let url = self.api_url(url);
-        let headers = self.auth_headers().await;
+        let headers = self.auth_headers().await?;
         Ok(self.get_http().post(&url, Some(&headers), payload).await?)
     }
 
@@ -136,7 +136,7 @@ where
     #[inline]
     async fn api_put(&self, url: &str, payload: &Value) -> ClientResult<String> {
         let url = self.api_url(url);
-        let headers = self.auth_headers().await;
+        let headers = self.auth_headers().await?;
         Ok(self.get_http().put(&url, Some(&headers), payload).await?)
     }
 
@@ -146,7 +146,7 @@ where
     #[inline]
     async fn api_delete(&self, url: &str, payload: &Value) -> ClientResult<String> {
         let url = self.api_url(url);
-        let headers = self.auth_headers().await;
+        let headers = self.auth_headers().await?;
         Ok(self
             .get_http()
             .delete(&url, Some(&headers), payload)
@@ -205,9 +205,15 @@ where
     /// - track_id - a spotify URI, URL or ID
     ///
     /// [Reference](https://developer.spotify.com/documentation/web-api/reference/#/operations/get-track)
-    async fn track(&self, track_id: TrackId<'_>) -> ClientResult<FullTrack> {
+    async fn track(
+        &self,
+        track_id: TrackId<'_>,
+        market: Option<Market>,
+    ) -> ClientResult<FullTrack> {
+        let params = build_map([("market", market.map(Into::into))]);
+
         let url = format!("tracks/{}", track_id.id());
-        let result = self.api_get(&url, &Query::new()).await?;
+        let result = self.api_get(&url, &params).await?;
         convert_result(&result)
     }
 
@@ -338,9 +344,9 @@ where
     async fn artist_top_tracks(
         &self,
         artist_id: ArtistId<'_>,
-        market: Market,
+        market: Option<Market>,
     ) -> ClientResult<Vec<FullTrack>> {
-        let params = build_map([("market", Some(market.into()))]);
+        let params = build_map([("market", market.map(Into::into))]);
 
         let url = format!("artists/{}/top-tracks", artist_id.id());
         let result = self.api_get(&url, &params).await?;
@@ -370,10 +376,15 @@ where
     /// - album_id - the album ID, URI or URL
     ///
     /// [Reference](https://developer.spotify.com/documentation/web-api/reference/#/operations/get-an-album)
-    async fn album(&self, album_id: AlbumId<'_>) -> ClientResult<FullAlbum> {
-        let url = format!("albums/{}", album_id.id());
+    async fn album(
+        &self,
+        album_id: AlbumId<'_>,
+        market: Option<Market>,
+    ) -> ClientResult<FullAlbum> {
+        let params = build_map([("market", market.map(Into::into))]);
 
-        let result = self.api_get(&url, &Query::new()).await?;
+        let url = format!("albums/{}", album_id.id());
+        let result = self.api_get(&url, &params).await?;
         convert_result(&result)
     }
 
@@ -386,10 +397,13 @@ where
     async fn albums<'a>(
         &self,
         album_ids: impl IntoIterator<Item = AlbumId<'a>> + Send + 'a,
+        market: Option<Market>,
     ) -> ClientResult<Vec<FullAlbum>> {
+        let params = build_map([("market", market.map(Into::into))]);
+
         let ids = join_ids(album_ids);
         let url = format!("albums/?ids={ids}");
-        let result = self.api_get(&url, &Query::new()).await?;
+        let result = self.api_get(&url, &params).await?;
         convert_result::<FullAlbums>(&result).map(|x| x.albums)
     }
 
@@ -495,11 +509,12 @@ where
     fn album_track<'a>(
         &'a self,
         album_id: AlbumId<'a>,
+        market: Option<Market>,
     ) -> Paginator<'_, ClientResult<SimplifiedTrack>> {
         paginate_with_ctx(
             (self, album_id),
             move |(slf, album_id), limit, offset| {
-                slf.album_track_manual(album_id.as_ref(), Some(limit), Some(offset))
+                slf.album_track_manual(album_id.as_ref(), market, Some(limit), Some(offset))
             },
             self.get_config().pagination_chunks,
         )
@@ -509,12 +524,17 @@ where
     async fn album_track_manual(
         &self,
         album_id: AlbumId<'_>,
+        market: Option<Market>,
         limit: Option<u32>,
         offset: Option<u32>,
     ) -> ClientResult<Page<SimplifiedTrack>> {
         let limit = limit.map(|s| s.to_string());
         let offset = offset.map(|s| s.to_string());
-        let params = build_map([("limit", limit.as_deref()), ("offset", offset.as_deref())]);
+        let params = build_map([
+            ("limit", limit.as_deref()),
+            ("offset", offset.as_deref()),
+            ("market", market.map(Into::into)),
+        ]);
 
         let url = format!("albums/{}/tracks", album_id.id());
         let result = self.api_get(&url, &params).await?;